Moved related methods inside Repo
