--- server/src/server_data/repos/connection.rs
@@ -1,20 +1,0 @@
-use diesel::pg::PgConnection;
-use diesel::prelude::*;
-use diesel::r2d2::{ConnectionManager, Pool, PoolError, PooledConnection};
-use dotenv::dotenv;
-use std::env;
-
-pub type PgPool = Pool<ConnectionManager<PgConnection>>;
-pub type PgPooledConnection = PooledConnection<ConnectionManager<PgConnection>>;
-
-fn init_pool(database_url: &str) -> Result<PgPool, PoolError> {
-    let manager = ConnectionManager::<PgConnection>::new(database_url);
-    Pool::builder().build(manager)
-}
-
-pub fn establish_connection() -> anyhow::Result<PgPool> {
-    dotenv().expect("Env file was not loaded correctly");
-    let database_url = env::var("DATABASE_URL")?;
-    PgConnection::establish(&database_url)?;
-    Ok(init_pool(&database_url)?)
-}
